+++ src/full/Agda/Interaction/Highlighting/Generate.hs
@@ -413,6 +413,7 @@ warningHighlighting' b w = case tcWarning w of
   TerminationIssue terrs     -> terminationErrorHighlighting terrs
   NotStrictlyPositive d ocs  -> positivityErrorHighlighting d ocs
   ConstructorDoesNotFitInData c s1 s2 err -> errorWarningHighlighting c
+  CoinductiveEtaRecord _x    -> deadcodeHighlighting w
   -- #3965 highlight each unreachable clause independently: they
   -- may be interleaved with actually reachable clauses!
   UnreachableClauses _ rs    -> foldMap deadcodeHighlighting rs